+++ src/pages/SignInPage.tsx
@@ -1,0 +1,13 @@
+import Header from '../components/Header';
+import SignIn from '../signIn/components/SignIn';
+
+const SignInPage = () => {
+  return (
+    <>
+      <Header />
+      <SignIn />
+    </>
+  );
+};
+
+export default SignInPage;
